Update mattsdesktoptest.css
Changes:
Nasir see that I have removed lines 16-18 this is because colour of text with in the header can be simplified to one line within the already provided footer.

What needs to be done:

Yes you have it staying to the page using absolute which is good!

I need this to be at the very bottom of the page. keep in mind that on this .css file that we have the body height set to 1080px try to do something reflecting that.